+++ tycho-core/src/main/java/org/eclipse/tycho/core/TargetPlatformConfiguration.java
@@ -43,6 +43,32 @@ public class TargetPlatformConfiguration implements DependencyResolverConfigurat
         first, minimal
     }
 
+    public enum LocalArtifactHandling {
+        /**
+         * local artifacts are included and override items from the target
+         */
+        include,
+        /**
+         * NOT SUPPORTED YET! local artifacts override items from the target that are
+         * <i>equivalent</i> (major and minor version must equal the version from the target).
+         */
+        equivalent,
+        /**
+         * NOT SUPPORTED YET! local artifacts override items from the target that are
+         * <i>compatible</i> (major version must equal the version from the target).
+         */
+        compatible,
+        /**
+         * NOT SUPPORTED YET! local artifacts override items from the target that match
+         * <i>exactly</i> the version from the target.
+         */
+        perfect,
+        /**
+         * local artifacts are ignored
+         */
+        ignore;
+    }
+
     private String resolver;
 
     private List<TargetEnvironment> environments = new ArrayList<>();
@@ -69,6 +95,8 @@ public class TargetPlatformConfiguration implements DependencyResolverConfigurat
     private Map<String, String> resolverProfileProperties = new HashMap<>();
 
     List<Supplier<File>> lazyTargetFiles = new ArrayList<>();
+
+    private LocalArtifactHandling localArtifactHandling;
 
     /**
      * Returns the list of configured target environments, or the running environment if no
@@ -224,6 +252,17 @@ public class TargetPlatformConfiguration implements DependencyResolverConfigurat
     @Override
     public Collection<IRequirement> getAdditionalRequirements() {
         return List.of();
+    }
+
+    public LocalArtifactHandling getIgnoreLocalArtifacts() {
+        if (localArtifactHandling == null) {
+            return LocalArtifactHandling.include;
+        }
+        return localArtifactHandling;
+    }
+
+    public void setLocalArtifactHandling(LocalArtifactHandling localArtifactHandling) {
+        this.localArtifactHandling = localArtifactHandling;
     }
 
 }