+++ tycho-extras/tycho-p2-extras-plugin/src/main/java/org/eclipse/tycho/plugins/p2/extras/CompareWithBaselineMojo.java
@@ -144,7 +144,7 @@ public class CompareWithBaselineMojo extends AbstractMojo {
         P2Resolver resolver = resolverFactory.createResolver(Collections.singletonList(runningEnvironment));
 
         TargetPlatformConfigurationStub baselineTPStub = new TargetPlatformConfigurationStub();
-        baselineTPStub.setForceIgnoreLocalArtifacts(true);
+        baselineTPStub.setIgnoreLocalArtifacts(true);
         baselineTPStub.setEnvironments(Collections.singletonList(runningEnvironment));
         for (String baselineRepo : this.baselines) {
             baselineTPStub.addP2Repository(toRepoURI(baselineRepo));